split components & swift formatted
